[feat] react form working
