+++ app/registration/Components/OldRegistration.css
@@ -1,0 +1,30 @@
+.make-offer-form {
+    max-width: 412px; /* Set the maximum width */
+    width: 100%; /* Make the element take up the full width of its parent */
+    /* margin: 24px;  */
+    display: flex;
+    flex-direction: column;
+    /* gap: 10px; */
+    align-self: center;
+    border: 1px solid black;
+    border-radius: 6px;
+    padding: 8px;
+  }
+  
+  .make-offer-wrapper {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    padding: 24px;
+  }
+  
+  .label-input-wrap {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    margin-bottom: 20px;
+  }
+  
+  form .account-button {
+    width: 100%;
+  }
