+++ package.json
@@ -26,6 +26,7 @@
     "postcss": "8.4.29",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-hook-form": "^7.46.2",
     "tailwindcss": "3.3.3",
     "typescript": "5.2.2",
     "uuid": "^9.0.1"
